requirements.txt
streamlit==1.45.0
google-analytics-data==0.18.15
google-auth==2.40.1
google-auth-oauthlib==1.2.1
pandas==2.2.3
plotly==6.0.1
